Improve navbar UI
