+++ src/components/NavBar/NavBar.module.css
@@ -1,0 +1,18 @@
+.navBarContainer {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 20px;
+  padding: 0 30px;
+}
+
+.navBarContainer .menu {
+  width: fit-content;
+}
+
+.navBarContainer h1 {
+  cursor: pointer;
+}
+.logo {
+  font-size: 90px;
+}
